+++ tests/_support/Helper/DBTestCase.php
@@ -5,6 +5,11 @@ namespace StellarWP\DB\Tests;
 use Codeception\TestCase\WPTestCase;
 use StellarWP\DB\DB;
 
+/**
+ * The base test case.
+ *
+ * @mixin \Codeception\PHPUnit\TestCase
+ */
 class DBTestCase extends WPTestCase {
 	protected $backupGlobals = false;